pyproject.toml
[project]
name = "gatepass"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.1.1",
]
